clean up and reformatting solutions

+++ 0001multiplesof3or5.py
@@ -7,4 +7,4 @@ mult3 = math.floor(999/3)
 mult5 = math.floor(999/5)
 mult15 = math.floor(999/15)
 
-print(triangular(mult3) * 3 + triangular(mult5) * 5 - triangular(mult15) * 15)
+print(int(triangular(mult3) * 3 + triangular(mult5) * 5 - triangular(mult15) * 15))
